Update index.html
We need to add a checkbox for night shifts that is stored as a boolean that will be checked by the java. 

if nights, then rate = rate + 2.50

There should be a subtitle too that points out the calculation does not work if you have overtime/holiday pay on your paystub. There is a way to do the calculation but it requires a lot more granular information that would make inputs cumbersome.